Breaking: drop `plot_` prefix from multiple functions (#189)
* breaking: drop plot_ prefix from plot functions

rationale: more consistent since not all functions had the prefix to begin with, should have been all or none. in the future, people might import pymatviz as pmv and call e.g. pmv.phonon_bands() in which case the plot_ feels redundant. so to reach consistency, seems better to remove plot_ prefix than to add it

* deprecate spacegroup_hist, still works but renamed spacegroup_bar in new bar.py module

* breaking: rename histogram(s->'').py